Switch to synchronous mkdirp.
Previously there was a bug for very small builds in that we could call mkdirp,
then minify a file and attempt to save it to cache before the directory was
created.  This switches to the synchronous version of mkdirp to prevent that.
This will negligibly impact performance as it is only called once per plugin,
and the call should only take milliseconds.