Separate cacheDir from other options passed to worker
This makes it so we pass a copy of the options object without the
`cacheDir` field to the worker. This way the worker only uses this
modified options object when creating cache keys for files. The
`cacheDir` value should generally not be factored into the cache key
derivation, and this specifically helps if the `cacheDir` value is not
consistent across runs.