Use uglifyJS sourceMap option instead of devtool.
This makes more sense as users may be directly using SourceMap plugins rather
than relying on the devtool setting. User's might also want to enable source
maps for some sections of their code but not others. This change enables them
to do that by using two instances of this plugin.