Major refactor away from child_process and to worker-farm.
Worker-farm provides a cleaner abstraction around managing child processes,
allowing us to call functions instead of manually passing messages.  This api
allows us to write less code, and the code that we do write is a bit cleaner.

Sadly some of the tests were removed as part of this refactor, but coverage
will be restored in the future.