Add uglify-es support.
uglify-es is a branch of uglify-js that is meant to enable uglifying of es6
code. This commit enables the newer uglify if the user provides uglifyES
options to the plugin.  I chose to support both since not everyone needs to
support older browsers, but those that do would likely appreciate if the tool
catches any sneaky es6.