Use babel-code-frame to generate pretty SyntaxError messages.
A developer on my team spent more time than needed debugging a SyntaxError
that was the result of attempting to minify es6 code that snuck into
node_modules/. I think that by providing additional context on the failed file,
we can help people be more productive.

This change uses babel-code-frame to generate an error message that includes a
few lines above/below the encountered SyntaxError.